pyproject.toml
[project]
name = "python-webscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "aiohttp>=3.12.13",
    "beautifulsoup4>=4.13.4",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "regex>=2024.11.6",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
